The Dashboard page makes the Experience component easier to read in the return statement
